backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
# Обновлено для совместимости с Python 3.13
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.12.1
psycopg2-binary==2.9.10

# Redis
redis==5.0.1
aioredis==2.0.1

# Celery
celery[redis]==5.4.0
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pydantic[email]==2.10.6

# TON SDK
pytonlib==0.0.53
tonsdk==1.0.13

# HTTP Client
httpx==0.27.2
aiohttp==3.9.0

# Validation & Serialization
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
pendulum==3.0.0
pytz==2023.3

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.2

# Dev Tools
ipython==8.18.0
watchfiles==0.21.0
